refactor: rename new screen in army selection screen
