--- modules/games/screens/New/index.ts
@@ -1,1 +1,0 @@
-export { default } from './New'
