--- modules/games/screens/New/ref.ts
@@ -1,6 +1,0 @@
-import { type BottomSheetModalMethods } from '@gorhom/bottom-sheet/lib/typescript/types'
-import { createRef } from 'react'
-
-const ref = createRef<BottomSheetModalMethods>()
-
-export default ref
