+++ modules/games/screens/index.ts
@@ -1,4 +1,4 @@
+export { default as ArmySelectionScreen } from './ArmySelection'
 export { default as GameScreen } from './Game'
 export { default as JoinGameScreen } from './Join'
-export { default as NewGameScreen } from './New'
 export { default as GamesTab } from './Tab'
